+++ Changelogs/Changelog.md
@@ -1,5 +1,7 @@
 ### Added
 - Browse now your previously selected achievements if you want to go back and forward with ease
+- Option to move the calendar button to open the achievement calendar
 
 ### Fixed
 - Errors occurring during combat should be drastically reduced when opening the achievements window
+- Strange behavior where more than one achievement is selected should be resolved